split database.types into two files
Otherwise the module will get too big.
