+++ src/ja/server/database/types/__init__.py
@@ -1,0 +1,4 @@
+"""
+This module contains several classes which represent the data stored in the
+database.
+"""
